Makes every collection operation work on `Coll a`
Some of the operations (head, tail, last, reduce, concat, map-concat) didn't
wrap the argument in a Coll type, as such `head` would not work if given a
single node.

This new design allows people to express things like:

    append (head a), b

Without having to make sure `a` is a collection first, which fits with the
nature of the `Coll` type and the overall design of the library.