Provides alternatives for (head foo xs)
Some functions are used too frequently to grab information on a specific node, making the user do (head <fun> (head xs)) all over the place. This is not only super tedious, but also leads to code duplication easily. Providing these functions out of the box is a better way to go.